+++ api/src/main/java/ca/bc/gov/educ/api/studentgraduation/config/GradStudentGraduationConfig.java
@@ -1,28 +1,34 @@
 package ca.bc.gov.educ.api.studentgraduation.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import ca.bc.gov.educ.api.studentgraduation.util.MessageHelper;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.netty.http.client.HttpClient;
 
 @Configuration
-@PropertySource("classpath:messages.properties")
-public class GradStudentGraduationConfig implements WebMvcConfigurer {
+public class GradStudentGraduationConfig {
 
-	@Autowired
-	RequestInterceptor requestInterceptor;
+	@Bean
+	public ModelMapper modelMapper() {
 
-	/**
-	 * Add interceptors.
-	 *
-	 * @param registry the registry
-	 */
-	@Override
-	public void addInterceptors(InterceptorRegistry registry) {
-		registry.addInterceptor(requestInterceptor).addPathPatterns("/**");
+		ModelMapper modelMapper = new ModelMapper();
+		//modelMapper.typeMap(GradProgramEntity.class, GradProgram.class);
+		//modelMapper.typeMap(GradProgram.class, GradProgramEntity.class);
+		return modelMapper;
+	}
+
+	@Bean
+	public WebClient webClient() {
+		HttpClient client = HttpClient.create();
+		client.warmup().block();
+		return WebClient.builder().build();
+	}
+
+	@Bean
+	public RestTemplate restTemplate(RestTemplateBuilder builder) {
+		return builder.build();
 	}
 }
