+++ api/src/main/java/ca/bc/gov/educ/api/studentgraduation/config/WebSecurityConfiguration.java
@@ -1,0 +1,38 @@
+package ca.bc.gov.educ.api.studentgraduation.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+public class WebSecurityConfiguration {
+
+    public WebSecurityConfiguration() {
+        SecurityContextHolder.setStrategyName(SecurityContextHolder.MODE_INHERITABLETHREADLOCAL);
+    }
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                .authorizeRequests()
+                .antMatchers(
+                        "/api/v1/api-docs-ui.html",
+                        "/api/v1/swagger-ui/**",
+                        "/api/v1/api-docs/**",
+                        "/actuator/health",
+                        "/actuator/prometheus",
+                        "/health")
+                .permitAll()
+                .anyRequest()
+                .authenticated()
+                .and()
+                .oauth2ResourceServer()
+                .jwt();
+        return http.build();
+    }
+
+}
